% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create.adpag.shell}
\alias{create.adpag.shell}
\title{Create a template shell script of postGGIR}
\usage{
create.adpag.shell()
}
\value{
The function will create a template shell script of postGGIR in the current directory, names as STUDYNAME_part0.maincall.R
}
\description{
Create a template shell script of postGGIR, named as STUDYNAME_part0.maincall.R.
}
